// client/packages/image-annotator/src/app.types.tsx
export type Maybe<T> = T | null | undefined;
// ...
export interface MetadataInfo {
    key: string;
    value: string;
}

export interface TooltipMetadata {
    [key: number]: MetadataInfo[];
}
// ...
export interface ToolSettingsSelect {
    tooltipOn: boolean;
}
// ...
export interface MouseEvent {
    point: paper.Point;
    item: paper.Item;
    modifiers?: {
        shift?: boolean;
    };
}